Hotfix for bad parse in docs example
One of the example queries demonstrates a bug in the Malloy grammar, specifically that two fields, one that is defined as an expression that is a function call, and another that is a field name reference, compiles as _two_ explore statements, rather than one. Adding a comma between the fields disambiguates.