remove the experimental ExploreDef implementation
There was a brief experiment with adding an "explore" along with "struct" as a thing which could have a name in the IR, which we ended up not using. In preparation for adding a "query" along with "struct" as thing which could have a name in the IR, it is time to remove "explore".