Allow import base url (#1342)
There are a few use cases, like notebook cells, where the document URL is not ideal for importing, but we still want to be able to import from. There's a bunch of hackery in the VS Code extension that makes this work, but it would be cleaner to push this into the parser itself. This adds an optional parser parameter, importBaseURL, that can be used as the basis for importing.