For nested fields we were using lateral flatten. The issue with it is that if the nested field is used in a join, snowflake complains that lateral flatten can't be on left side of join (https://stackoverflow.com/questions/63397022/snowflake-lateral-cannot-be-on-the-left-side-of-join)
fix is to always convert lateral flatten to a left join

Also simplify the array[scalar] vs array[record] treatment. earlier, at the
unnesting step we were creating a object for array[scalar] case to make it
consistent with the array[record] case. Now at the reference time, based on
the case, use the right expression
- array[record] : parent.value:sql_field
- array[scalar] : parent.value

Signed-off-by: Amit Aggarwal <amit@datairis.io>