Update PR tests to use pull_request_target (#2087)
This grants external PRs access to secrets in order to be able to run database tests. For security purposes external PRs will not automatically run tests, tests must be triggered by someone with write permissions to the repo (the core Malloy team).

This changes the names of the tests that are run, so older PRs will need to be updated to include these changes in order to pass the required test checks.